refactor(desktop): split mastra chat message components
