+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/ChatMastraPane/ChatMastraInterface/components/ChatMastraMessageList/components/AssistantMessage/index.ts
@@ -1,0 +1,1 @@
+export { AssistantMessage } from "./AssistantMessage";
